refactor(ui): вынести вкладки редактора документа по файлам (#490)
Файл на 1220 строк был оболочкой вкладок и двумя самими вкладками:
RequisitesTab 604 строки, GenerationTab 212, InstanceEditor 216.

Резал по тому же принципу, что и #488: перенос, а не разбор. Внутри
RequisitesTab швы есть — «Основа» (#71), рендерер ячейки, рейл разделов, — но
крупнейший из них, renderCell на ~150 строк, замыкает ЧЕТЫРНАДЦАТЬ переменных:
вынести его компонентом значит завести интерфейс на четырнадцать пропсов, а это
хуже длинной функции, а не лучше. Настоящая проблема была не в длине компонента,
а в том, что оболочка и вкладки лежали вместе.

index.tsx 1220 → 280, вкладки по своим файлам.

containsRef и BrokenCountBadge вынесены отдельными модулями, а не оставлены в
index: импорт из index создал бы цикл, ведь index импортирует вкладки. Чистая
функция и компонент разведены по разным файлам — иначе справедливо ругается
react-refresh/only-export-components. Тип SaveRef объявлен во вкладке, которая
его и определяет.

Попутно снят дублирующий разделитель в ConstructionsList — по замечанию ревью
предыдущего PR.

Проверено живьём: редактор открыт по прямой ссылке ?doc=, «Реквизиты»
отрисованы полностью (рейл разделов со счётчиками, поля всех видов, таблицы,
строка-подсказка «Основа не выбрана»), «Генерация» показывает шаблоны и кнопки,
переключение туда-обратно живо, ошибок страницы нет. tsc чист, 212 тестов
зелёные, ошибок eslint ровно три — все перенесённые, столько же было на master.
